refactor: 코드 리팩터링 및 새로고침 시 서버 지연시간으로 인해 username이 깜빡이던 현상 해결
