+++ src/components/header/header.html
@@ -11,7 +11,9 @@
   <div class="header-menu">
     <ul class="menu">
       <li class="code-btn">
-        <button class="create-code-btn hidden">코드 작성</button>
+        <a href="/write">
+          <button class="create-code-btn hidden">코드 작성</button>
+        </a>
       </li>
       <li class="user-name hidden">
         <div class="user-name-text"></div>
