+++ src/components/header/header.js
@@ -1,5 +1,3 @@
 import headerHTML from "./header.html?raw";
 
-document.addEventListener("DOMContentLoaded", () => {
-  document.body.insertAdjacentHTML("afterbegin", headerHTML);
-});
+document.body.insertAdjacentHTML("afterbegin", headerHTML);
