+++ src/components/login-modal/login-modal.html
@@ -1,0 +1,103 @@
+<div id="login-modal" class="modal-overlay hidden">
+  <div class="modal-content">
+    <button class="modal-close-btn">&times;</button>
+    <img src="/sh4reLogo.svg" class="login-sh4re-icon" />
+    <form id="login-form" class="login-form">
+      <div class="form-group">
+        <label for="login-username" class="visually-hidden">아이디</label>
+        <input
+                placeholder="아이디"
+                type="text"
+                id="login-username"
+                name="username"
+                class="text-box"
+                required
+        />
+      </div>
+      <div class="form-group">
+        <label for="login-password" class="visually-hidden">비밀번호</label>
+        <input
+                placeholder="비밀번호"
+                type="password"
+                id="login-password"
+                name="password"
+                class="text-box"
+                autocomplete="off"
+                required
+        />
+      </div>
+      <button type="submit" class="login-submit-btn">로그인</button>
+    </form>
+    <form id="signIn-form" class="signup-form hidden-signup">
+      <div class="form-group">
+        <label for="name" class="visually-hidden">이름</label>
+        <input
+                placeholder="이름"
+                type="text"
+                id="name"
+                name="name"
+                class="text-box"
+                required
+        />
+      </div>
+      <div id="signup-form">
+        <div class="student-number-box">
+          <label for="grade" class="visually-hidden">학년</label>
+          <select name="grade" id="grade" class="student-number">
+            <option value="default" disabled selected>학년</option>
+            <option value="1">1학년</option>
+            <option value="2">2학년</option>
+            <option value="3">3학년</option>
+          </select>
+          <label for="class" class="visually-hidden">반</label>
+          <select name="classNumber" id="class" class="student-number">
+            <option value="default" disabled selected>반</option>
+            <option value="1">1반</option>
+            <option value="2">2반</option>
+            <option value="3">3반</option>
+            <option value="4">4반</option>
+          </select>
+          <label for="number" class="visually-hidden">번호</label>
+          <input
+                  placeholder="번호"
+                  type="text"
+                  id="number"
+                  name="studentNumber"
+                  min="0"
+                  max="19"
+                  pattern="^(?:[1-9]|1[0-9])$"
+                  class="student-number"
+                  required
+          />
+        </div>
+      </div>
+      <div class="form-group">
+        <label for="username" class="visually-hidden">아이디</label>
+        <input
+                placeholder="아이디"
+                type="text"
+                id="signup-username"
+                name="username"
+                class="text-box"
+                required
+        />
+      </div>
+      <div class="form-group">
+        <label for="password" class="visually-hidden">비밀번호</label>
+        <input
+                placeholder="비밀번호"
+                type="password"
+                id="signup-password"
+                name="password"
+                class="text-box"
+                autocomplete="off"
+                required
+        />
+      </div>
+      <button type="submit" class="login-submit-btn">회원가입</button>
+    </form>
+    <div class="modal-options">
+      <a class="create-account">아직 계정이 없으신가요?</a>
+    </div>
+  </div>
+</div>
