+++ src/components/login-modal/login-modal.js
@@ -1,0 +1,3 @@
+import loginModal from "./login-modal.html?raw";
+
+document.body.insertAdjacentHTML("afterbegin", loginModal);
