+++ src/js/main.js
@@ -2,4 +2,6 @@ import "/styles/style.scss";
 import "/styles/signUp.scss";
 import "/styles/login.scss";
 import "/components/header/header.js";
-import { customFetch } from "./common";
+import "/components/login-modal/login-modal.js";
+import "./auth.js"
+import "./highlight.js"
