--- src/js/write.js
@@ -1,9 +1,0 @@
-document.addEventListener("DOMContentLoaded", () => {
-  const createCodeBtn = document.querySelector(".create-code-btn");
-
-  if (createCodeBtn) {
-    createCodeBtn.addEventListener("click", () => {
-      window.location.href = "write"; // write.html로 이동
-    });
-  }
-});
