+++ src/write.html
@@ -26,7 +26,10 @@
         <button class="share">코드 공유</button>
       </div>
     </main>
+    <footer id="page-footer">
+      <p>&copy; 2025 Team-Sh4re</p>
+    </footer>
     <script type="module" src="/js/main.js"></script>
-    <script type="module" src="/js/login.js"></script>
+    <script type="module" src="/js/write.js"></script>
   </body>
 </html>
